Modified the Event/Case models to add sublists.
Added more getter methods to EventType (so that we
can get victims, investigators and perpetrators easily).

Added to Event, Now contains lists of victims,
investigators and perpetrators. These properties are
sub-sets of people and read only.

Added to CaseFile, a way of getting the sublists added
to Event

Added to the seeds "Investigator"