Many new files are due to upgrading to rails edge (really wanted to use enum_prefix, despite this soon changing to _prefix, which i'll bug in a minute.
Added a new migration which adds a large number of fields to
people. This commit is the first stage of making that much more
complex.

Changed the code in the controller to allow the new fields to be
recieved.

Added enums to person model for new fields

Modified the person form to include these fields. Only the new/edit
form will currently contain these. They are not included in show yet.