Fixes #87
People now have aliases. This is a major change to how the system
works but I've tried to make the transition as seemless as possible
with the migrations.

People still have names. This is handled by having a "default alias"
behind the scenes. Setting the persons name will handle this
automatically.

Names can also be added/removed on the edit person screen.